+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Change Log - @itwin/connector-framework
 
+## 2.1.0-dev.2
+
+- move access token handling to base connector
+- check expiration before returning a cached token
+
 ## 2.1.0-dev.1
+
 - add support for channel keys
 - add connector's channel to key to allowed channels after briefcase is loaded
 - make the connector's job subject the channel root